tests/wpt/web-platform-tests/fullscreen/api/element-request-fullscreen-null-ns-manual.html

<!DOCTYPE html>
<title>Element#requestFullscreen() for an element in null namespace</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  const element = document.createElementNS(null, 'a');
  assert_equals(Object.getPrototypeOf(element), Element.prototype);
  document.body.appendChild(element);
  trusted_request(t, element);
  document.onfullscreenchange = t.unreached_func("fullscreenchange event");
  document.onfullscreenerror = t.step_func_done();
});
</script>
